requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Werkzeug==2.3.2
PyJWT==2.8.0
